+++ CHANGELOG.md
@@ -8,6 +8,7 @@ using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 ### Added
 
 - Added new **JSON** reporter that creates a JSON file with all validation errors.
+- Add option to register **custom translator services* by using the TranslatorFactory **register** command.
 
 ## [1.11.0]